template: prefer "history" to "versions"
When viewing a templates versions, the heading copy implied that you
were only looking at previous versions of the template when in fact the
topmost version was also the latest.

By using "history of changes" instead, it should be clearer that you're
looking at the full history of changes to the template.