api/rate-limits: count aggregate notifications
across type, status, for the current day.

Results could be inconsistent depending on whether there were multiple
notification statuses out at the time of check, or even if there are
multiple templates per service.

This is a bug that is especially prominent when redis is disabled.

originally surfaced here:
  https://github.com/alphagov/notifications-api/commit/e3d418506a072fa8802800b5575719e3d097a5b1

This was introduced through a refactor to make this check more efficient --
Originally there was a helper function that collected some stats to be
displayed somewhere. This helper was also being used here to seed message
limit counts by totalling all the stats. When this was refactored it was
mistakenly decided to keep the "group by" clause BUT also to stop adding up
all the groups of stats.

We could have reverted to the old behaviour and added stats together
again, or just stop grouping.