admin: force user log out when signing in again
We came across an issue with one session that was possibly corrupted. It
resulted in infinite redirects to the /sign-in page. This seems to have
been caused by the sign in route thinking that there is a current_user
and that the current_user.is_authenticated, however the Flask
login_required decorator disagreed with this state. This caused the
Flask decorator to redirect to the login_manager.unauthorized_handler
handler, which in turn would redirect back to /sign-in. This caused the
infinite redirect.

To fix this, we ensure that logout_user is called in the
unauthorized_handler. This should be OK since the user should already be
logged out if we enter this handler.