admin: fix infinite redirect when session invalidated
This reverts commit 1b7eaa27ae3a6a73f8aa8b41e1baae2dd5a39d5b.

To replicate the issue follow these steps:

1. Sign in to a browser window (window 1)
2. Visit a service page, e.g. /services/1234
3. Copy the current URL.
4. Open a private browser window, pasting in the URL from step 3 (window 2).
5. You'll be redirected to /sign-in, so sign in using the same user credentials.
6. You'll then be redirected to /services/1234 and can successfully see the service page.
7. Go back to window 1 and refresh.
8. Notice that the window redirects infinitely to /sign-in.

This seems to have been caused by the sign in route thinking that there is a current_user and that the current_user.is_authenticated, however the Flask login_required decorator disagreed with this state. This caused the Flask decorator to redirect to the login_manager.unauthorized_handler handler, which in turn would redirect back to /sign-in. This caused the infinite redirect.