admin: add pages to find and archive users
This brings in the bare minimum needed to wire up these pages. The
admin/app/models/*.py files are not all required, but it was easier to
just bring in the entire folder and just disable some parts of it that
imported more things (that would have caused cascading imports).