Refactor sending to international phone numbers
Our two SMS providers expect the number to be given in E.164 format. We've refactored some things so that we can pass around the number already in E.164 format. For new notification objects we now set the normalised_to field to the E.164 format and use that throughout the codebase instead of validating and formatting on the fly.