admin: update gulp-base64, pin Node version
gulp-base64 was tagged a while back, but we didn't update package.json
to pull in that change.

It was discovered that the Node frontend will not work on 13.x so this
commit also pins it to 10.x. In order for a 10.x to be pulled in by
CircleCI, the circleci/python image needed to be updated to 3.6.8-node.

Also tidy up some fields in package.json.