Revert "celery: use SQS backend"
This reverts commit 6f171b907b99429be6ae4dc43184dda2b2fd09fe.

A later commit will reintroduce these changes in a way that is more
backwards compatible with prod. The idea will be to keep the current
worker apps running so that they can drain the tasks in the DB. The API
will start putting new messages onto the SQS queues. And the new workers
will start processing messages off the SQS queues.