celery: use SQS backend
In order to scale the delivery of messages, the celery worker is split off into two CF apps. One app processes the `send-sms-tasks` and `send-email-tasks` queues. The other app processes all other queues. This change also allows for the delivery worker to be scaled up if necessary. A later change to consider would be to split off the `database-tasks` to be serviced by its own worker CF app. Since these tasks are database heavy, it might be worth being able to scale these as well.

In order to keep the celery DB tasks processing in prod we now use a different worker app name for the SQS workers. This means that the old workers can continue to drain the queue of any existing tasks.

This change also required upgrading Celery to 4.4.0.

Also in this commit:

- switch to using a dedicated RDS instance in feature environments. A drawback of this is that it takes 15-20minutes to provision the database. This only applies on the first push of the feature branch and subsequent CI builds will be much faster. The tradeoff here is probably worth it because it makes the feature environments more like production. This is useful for testing a feature environment, especially when testing production loads.
- wait for postgres DB service to be created
- fix STG variable prefix consistency in ci/Makefile
- add debug-notification command which is useful to find out state information about a given notification