sap: add new provider (inbound, outbound)
Also in this commit:

- delete old providers
- remove dead MMG and Firetext code
- add inbound code for Twilio
- command: allow inserting one number, not from file because we don't have access to the CSV file
- docs: update for inbound sms callbacks
- docs: only show languages that have examples
- docs: add delivery callback example

Also remove custom handling for 'send-sms-code' reference. This looks
like it was part of some old UK code when the system only sent out SMS
codes and verify emails. This commit shows the reference being removed
from the callers, but it looks like it was never removed in the tests
and in the processing function.